+++ R/consolidate.R
@@ -120,8 +120,11 @@ consolidate <- function(files = NULL, path = NULL, dest.path = NULL, pb = TRUE, 
   }
 
   # stop if files are not in working directory
-  if (length(files) == 0) stop2("no files found in working directory and/or subdirectories") else message2(paste(length(files), "files were found"))
-
+  if (length(files) == 0) stop2("no files found in working directory and/or subdirectories") else message2(paste(length(files), "files were found\n"))
+  
+  # collect file metadata
+  message2("Collecting file metadata (can take some time)\n")
+  
   # create new names for duplicated songs
   old_name <- basename(files)
   files <- files[order(old_name)]